requirements.txt
fastapi
uvicorn
sse-starlette
python-chess
qdrant-client
numpy
tqdm
